pyproject.toml
[project]
name = "llm-benchmarking"
version = "0.1.0"
description = "Modular LLM benchmark tracker with Ollama analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.2.0",
    "beautifulsoup4",
    "httpx",
    "lxml",
    "langchain",
    "langchain-ollama",
    "pandas>=2.1.0",
    "playwright",
    "plotly>=5.18.0",
    "pydantic",
    "pyyaml",
    "rich",
    "streamlit>=1.30.0",
    "typer",
]

[project.scripts]
llm-bench = "llm_benchmarks.main:app"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_benchmarks"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
